requirements.txt
pandas
streamlit
python-dotenv
tabulate
google-generativeai
matplotlib
seaborn
markdown
pdfkit
jinja2
